refactor: 💡 more cleanups
