+++ server/src/internal/api/rewards/handlers/rewardPrograms/handleCreateTrigger.ts
