+++ server/src/internal/api/rewards/handlers/rewardPrograms/handleDeleteRewardProgram.ts
