+++ server/src/internal/api/rewards/rewardProgramRouter.ts
@@ -1,5 +1,5 @@
 import express, { type Router } from "express";
-import { handleCreateRewardProgram, handleDeleteRewardProgram } from "./handlers/referrals/index.js";
+import { handleCreateRewardProgram, handleDeleteRewardProgram } from "./handlers/rewardPrograms/index.js";
 
 export const rewardProgramRouter: Router = express.Router();
